refactor: Move shared code between PlayerPageSimulation3D and InventorySimulation3D into PlayerSimulation3D
